chore(components): reorganize components and stories
