+++ packages/components/src/components/Modal/Modal.tsx
@@ -4,11 +4,11 @@ import styled, { css } from 'styled-components';
 
 import { useKeyPress } from '@trezor/react-utils';
 import { useTheme } from '../../utils/hooks';
-import { Icon } from '../Icon/Icon';
+import { Icon } from '../assets/Icon/Icon';
 import { H1 } from '../typography/Heading/Heading';
 import { variables } from '../../config';
 import { IconType } from '../../support/types';
-import { Progress } from '../loaders/Progress';
+import { Progress } from '../loaders/Progress/Progress';
 
 const CLOSE_ICON_SIDE = 26;
 const CLOSE_ICON_PADDING = 16;
