+++ packages/components/src/components/Tooltip/all.stories.tsx
@@ -15,7 +15,7 @@ const RichContentExample = styled.div`
     text-align: left;
 `;
 
-storiesOf('Tooltip', module).add(
+storiesOf('Misc/Tooltip', module).add(
     'All',
     () => (
         <>
