+++ packages/components/src/components/Warning/Warning.stories.tsx
@@ -1,40 +1,18 @@
-import { StoryObj } from '@storybook/react';
-
-import { Warning } from '../../index';
+import { Warning as WarningComponent } from '../../index';
 
 export default {
     title: 'Misc/Warning',
-    component: Warning,
+    component: WarningComponent,
 };
 
-export const Basic: StoryObj<typeof Warning> = {
+export const Warning = {
     args: {
-        children: 'Warning! Here dragons abound. 🐲',
-        variant: 'warning',
+        children: 'Insert text here.',
         withIcon: true,
     },
-};
-
-export const Critical: StoryObj<typeof Warning> = {
-    args: {
-        children: 'Critical! Dragons! Run!',
-        variant: 'critical',
-        withIcon: true,
-    },
-};
-
-export const Info: StoryObj<typeof Warning> = {
-    args: {
-        children: 'Info: Dragons have been spotted in this area.',
-        variant: 'info',
-        withIcon: true,
-    },
-};
-
-export const Learn: StoryObj<typeof Warning> = {
-    args: {
-        children: 'Learn: This variant is used in Guide.',
-        variant: 'learn',
-        withIcon: true,
+    argTypes: {
+        className: {
+            control: false,
+        },
     },
 };
