+++ packages/components/src/components/Warning/Warning.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { transparentize } from 'polished';
 
-import { Icon } from '../Icon/Icon';
+import { Icon } from '../assets/Icon/Icon';
 import { useTheme } from '../../utils';
 import { variables } from '../../config';
