+++ packages/components/src/components/assets/CoinLogo/CoinLogo.tsx
