+++ packages/components/src/components/assets/CoinLogo/coins.ts
