+++ packages/components/src/components/assets/Flag/Flag.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import React from 'react';
-import { FlagType } from '../../support/types';
+import { FlagType } from '../../../support/types';
 
 const Wrapper = styled.div`
     display: flex;
@@ -17,7 +17,7 @@ export const Flag = ({ size = 24, country, className }: FlagProps) => (
     <Wrapper>
         <img
             // eslint-disable-next-line global-require, import/no-dynamic-require
-            src={require(`../../images/flags/${country.toLowerCase()}.svg`)}
+            src={require(`../../../images/flags/${country.toLowerCase()}.svg`)}
             width={`${size}px`}
             alt={`flag-${country}`}
             className={className}
