+++ packages/components/src/components/assets/Icon/Icon.stories.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { Icon, variables, colors } from '../../index';
+import { Icon, variables, colors } from '../../../index';
 import { storiesOf } from '@storybook/react';
 import { number, select, color } from '@storybook/addon-knobs';
 
-storiesOf('Icons', module).add('Icon', () => {
+storiesOf('Assets/Icons', module).add('Icon', () => {
     const size = number('Size', 24);
     const iconColor = color('Color', colors.TYPE_LIGHT_GREY);
