+++ packages/components/src/components/assets/Icon/icons.ts
@@ -1,0 +1,115 @@
+/* eslint-disable global-require */
+export const ICONS = {
+    ANCHOR: require('../../../images/icons/anchor.svg'),
+    APP: require('../../../images/icons/app.svg'),
+    ARROW_DOWN: require('../../../images/icons/arrow_down.svg'),
+    ARROW_LEFT_LONG: require('../../../images/icons/arrow_left_long.svg'),
+    ARROW_LEFT: require('../../../images/icons/arrow_left.svg'),
+    ARROW_RIGHT_LONG: require('../../../images/icons/arrow_right_long.svg'),
+    ARROW_RIGHT: require('../../../images/icons/arrow_right.svg'),
+    ARROW_UP: require('../../../images/icons/arrow_up.svg'),
+    ARTICLE: require('../../../images/icons/article.svg'),
+    ASTERISK: require('../../../images/icons/asterisk.svg'),
+    BACK: require('../../../images/icons/back.svg'),
+    BACKEND: require('../../../images/icons/backend.svg'),
+    BACKUP: require('../../../images/icons/backup.svg'),
+    BINARY: require('../../../images/icons/binary.svg'),
+    BLOCKED: require('../../../images/icons/blocked.svg'),
+    BROADCAST: require('../../../images/icons/broadcast.svg'),
+    CALENDAR: require('../../../images/icons/calendar.svg'),
+    CANCEL: require('../../../images/icons/cancel.svg'),
+    CHANGE_ADDRESS: require('../../../images/icons/change_address.svg'),
+    CHECK_ACTIVE: require('../../../images/icons/check_active.svg'),
+    CHECK: require('../../../images/icons/check.svg'),
+    CLOCK_ACTIVE: require('../../../images/icons/clock_active.svg'),
+    CLOCK: require('../../../images/icons/clock.svg'),
+    COIN_CONTROL: require('../../../images/icons/coin_control.svg'),
+    COIN: require('../../../images/icons/coin.svg'),
+    COINS: require('../../../images/icons/coins.svg'),
+    CONFETTI_SUCCESS: require('../../../images/icons/confetti.svg'),
+    COOKIE: require('../../../images/icons/cookie.svg'),
+    COPY: require('../../../images/icons/copy.svg'),
+    CROSS: require('../../../images/icons/cross.svg'),
+    CSV: require('../../../images/icons/csv.svg'),
+    CUBE: require('../../../images/icons/cube.svg'),
+    DASHBOARD: require('../../../images/icons/dashboard.svg'),
+    DATA: require('../../../images/icons/data.svg'),
+    DISCREET: require('../../../images/icons/discreet.svg'),
+    DOWN: require('../../../images/icons/down.svg'),
+    DROPBOX: require('../../../images/icons/dropbox.svg'),
+    EJECT: require('../../../images/icons/eject.svg'),
+    EXPERIMENTAL: require('../../../images/icons/experimental.svg'),
+    EXTERNAL_LINK: require('../../../images/icons/external_link.svg'),
+    FEEDBACK: require('../../../images/icons/feedback.svg'),
+    FINGERPRINT: require('../../../images/icons/fingerprint.svg'),
+    FIRMWARE: require('../../../images/icons/firmware.svg'),
+    FLAG: require('../../../images/icons/flag.svg'),
+    GAS: require('../../../images/icons/gas.svg'),
+    GHOST: require('../../../images/icons/ghost.svg'),
+    GITHUB: require('../../../images/icons/github.svg'),
+    GOOGLE_DRIVE: require('../../../images/icons/google_drive.svg'),
+    GRID: require('../../../images/icons/grid.svg'),
+    HIDE: require('../../../images/icons/hide.svg'),
+    HOLOGRAM: require('../../../images/icons/hologram.svg'),
+    INFO: require('../../../images/icons/info.svg'),
+    KEY: require('../../../images/icons/key.svg'),
+    LIGHTBULB: require('../../../images/icons/lightbulb.svg'),
+    LOCK_ACTIVE: require('../../../images/icons/lock_active.svg'),
+    LOCK: require('../../../images/icons/lock.svg'),
+    LOG: require('../../../images/icons/log.svg'),
+    MENU: require('../../../images/icons/menu.svg'),
+    MORE: require('../../../images/icons/more.svg'),
+    NEW: require('../../../images/icons/new.svg'),
+    NEWSPAPER: require('../../../images/icons/newspaper.svg'),
+    NON_SUITE: require('../../../images/icons/non_suite.svg'),
+    NOTIFICATION: require('../../../images/icons/notification.svg'),
+    OS_LINUX_ARM64: require('../../../images/icons/os_linux_arm64.svg'),
+    OS_LINUX: require('../../../images/icons/os_linux.svg'),
+    OS_MAC: require('../../../images/icons/os_mac.svg'),
+    OS_WINDOWS: require('../../../images/icons/os_windows.svg'),
+    PACKAGE: require('../../../images/icons/package.svg'),
+    PALETTE: require('../../../images/icons/palette.svg'),
+    PAUSE: require('../../../images/icons/pause.svg'),
+    PENCIL_LINE: require('../../../images/icons/pencil_line.svg'),
+    PENCIL: require('../../../images/icons/pencil.svg'),
+    PIN: require('../../../images/icons/pin.svg'),
+    PLAY: require('../../../images/icons/play.svg'),
+    PLUS: require('../../../images/icons/plus.svg'),
+    QR: require('../../../images/icons/qr.svg'),
+    QUESTION: require('../../../images/icons/question.svg'),
+    RBF: require('../../../images/icons/rbf.svg'),
+    RECEIVE: require('../../../images/icons/receive.svg'),
+    RECOVER: require('../../../images/icons/recover.svg'),
+    REFRESH: require('../../../images/icons/refresh.svg'),
+    SEARCH: require('../../../images/icons/search.svg'),
+    SEED_SHAMIR: require('../../../images/icons/seed_shamir.svg'),
+    SEED_SINGLE: require('../../../images/icons/seed_single.svg'),
+    SEND: require('../../../images/icons/send.svg'),
+    SETTINGS: require('../../../images/icons/settings.svg'),
+    SHIELD_CHECK: require('../../../images/icons/shield_check.svg'),
+    SHIELD_CROSS: require('../../../images/icons/shield_cross.svg'),
+    SHOW: require('../../../images/icons/show.svg'),
+    SHUFFLE: require('../../../images/icons/shuffle.svg'),
+    SIGNATURE: require('../../../images/icons/signature.svg'),
+    STOP: require('../../../images/icons/stop.svg'),
+    TREZOR_T1B1: require('../../../images/icons/trezor_t1b1.svg'),
+    TREZOR_T2T1: require('../../../images/icons/trezor_t2t1.svg'),
+    TREZOR_T2B1: require('../../../images/icons/trezor_t2b1.svg'),
+    TABLE: require('../../../images/icons/table.svg'),
+    TAG_MINIMAL: require('../../../images/icons/tag_minimal.svg'),
+    TAG: require('../../../images/icons/tag.svg'),
+    TOR_MINIMAL: require('../../../images/icons/tor_minimal.svg'),
+    TOR: require('../../../images/icons/tor.svg'),
+    TRANSFER: require('../../../images/icons/transfer.svg'),
+    TREZOR_LOGO: require('../../../images/icons/trezor_logo.svg'),
+    UP: require('../../../images/icons/up.svg'),
+    USERS: require('../../../images/icons/users.svg'),
+    VERIFIED: require('../../../images/icons/verified.svg'),
+    WALLET_HIDDEN: require('../../../images/icons/wallet_hidden.svg'),
+    WALLET: require('../../../images/icons/wallet.svg'),
+    WARNING: require('../../../images/icons/warning.svg'),
+    ONE_USER: require('../../../images/icons/one_user.svg'),
+    TWO_USERS: require('../../../images/icons/two_users.svg'),
+    THREE_USERS: require('../../../images/icons/three_users.svg'),
+    FOUR_USERS: require('../../../images/icons/four_users.svg'),
+};
