+++ packages/components/src/components/assets/TrezorLogo/trezorLogos.ts
