+++ packages/components/src/components/loaders/Progress/Progress.stories.tsx
