+++ packages/components/src/index.ts
@@ -12,7 +12,7 @@ export * from './utils/hooks';
 export * from './components/buttons/Button/Button';
 export * from './components/buttons/PinButton/PinButton';
 export * from './components/buttons/TooltipButton/TooltipButton';
-export * from './components/Flag/Flag';
+export * from './components/assets/Flag/Flag';
 export * from './components/Dropdown/Dropdown';
 export * from './components/form/Input/Input';
 export * from './components/form/Textarea/Textarea';
@@ -22,7 +22,7 @@ export * from './components/form/RadioButton/RadioButton';
 export * from './components/form/Range/Range';
 export * from './components/form/Switch/Switch';
 export * from './components/form/InputStyles';
-export * from './components/Icon/Icon';
+export * from './components/assets/Icon/Icon';
 export * from './components/typography/Heading/Heading';
 export * from './components/typography/Paragraph/Paragraph';
 export * from './components/typography/Link/Link';
@@ -32,11 +32,11 @@ export * from './components/Card/Card';
 export * from './components/ConfirmOnDevice/ConfirmOnDevice';
 export * from './components/Modal/Modal';
 export * from './components/Modal/Backdrop';
-export * from './components/logos/CoinLogo/CoinLogo';
-export * from './components/logos/TrezorLogo/TrezorLogo';
-export * from './components/loaders/Spinner';
-export * from './components/loaders/Progress';
-export * from './components/loaders/LoadingContent';
+export * from './components/assets/CoinLogo/CoinLogo';
+export * from './components/assets/TrezorLogo/TrezorLogo';
+export * from './components/loaders/Spinner/Spinner';
+export * from './components/loaders/Progress/Progress';
+export * from './components/loaders/LoadingContent/LoadingContent';
 export * from './components/Tooltip/Tooltip';
 export * from './components/Timerange/Timerange';
 export * from './components/Truncate/Truncate';
